new update

+++ database/migrations/2019_07_08_103240_create_invoices_table.php
@@ -21,7 +21,7 @@ class CreateInvoicesTable extends Migration
             $table->date('date');
             $table->string('invoice_number');
             $table->integer('quantity');
-            $table->integer('select_point');
+            $table->float('select_point');
             $table->integer('point');
             $table->string('kyat');
             $table->string('pal');
